requirements.txt
Flask==2.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
numpy==1.23.0
pandas==1.4.3
psycopg2-binary==2.9.3
scikit-learn==1.0.1
scipy==1.8.1
threadpoolctl==3.1.0
Werkzeug==2.1.2
zipp==3.8.0
